Refactor: 코드 리뷰 반영 - 엔티티 및 DTO 수정

+++ src/main/java/com/bootcamp/paymentdemo/order/dto/response/OrderItemResponse.java
@@ -12,7 +12,7 @@ public record OrderItemResponse(
         return new OrderItemResponse(
                 item.getProductId(),
                 item.getProductName(),
-                (long) item.getPrice(),
+                item.getPrice(),
                 item.getQuantity()
         );
     }
